+++ src/main/java/calculator/Number.java
@@ -1,4 +1,4 @@
-package step2;
+package calculator;
 
 import java.util.Objects;
