+++ src/main/java/lotto/application/io/Input.java
@@ -1,4 +1,4 @@
-package step3.application.io;
+package lotto.application.io;
 
 import static java.util.stream.Collectors.toList;
 
@@ -6,7 +6,7 @@ import java.io.Closeable;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
-import step3.domain.LottoNumber;
+import lotto.domain.LottoNumber;
 
 public class Input implements Closeable {
