+++ src/main/java/lotto/domain/CreateLottoNumberPolicy.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package lotto.domain;
 
 import static java.util.stream.Collectors.toList;
