+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package lotto.domain;
 
 import static java.util.stream.Collectors.toList;
