+++ src/main/java/lotto/domain/Rank.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package lotto.domain;
 
 import java.util.Arrays;
