+++ src/test/java/lotto/domain/StatisticsTests.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
